Add batch support to `rgb_to_lab`, `lab_to_rbg`, and `hsv_to_rbg`
We are refactoring the image processing methods for CURL so that they
support batches. For the LAB conversion methods the main refactors were:

- Using einsum to do batch matrix multiplication, instead of using
matmul for a single multiplication.
- Reshaping 1D tensors to broadcast them correctly when adding or
multiplying them to 4D image batch tensors.
- Eliminate the permute operation, as this makes the operation more
confusing (unintuitive tensor shapes) and its only purpose was to make
broadcasting possible (reshaping is a better solution).

For the hsv_to_rgb method we:

- Eliminated the permutes (seemed to serve no purpose here).
- Changed indexing operations in the operations so that the right
channels were pulled out for each of the formulas.

We suspect this method was structured in this way just to match the LAB
methods, not for any functional reason.

We also added comments walking through the `rgb_to_hsv` method so we can
understand it when we come back to it to refactor it.