Batch support for `rgb_to_hsv` and apply curve methods
We modified the methods rgb_to_hsv, apply_curve, adjust_hsv, adjust_rgb,
and adjust_lab to support batches. For the first we eliminated
unnecessary permutation, chunking, etc. operations and worked with the
input image batch tensor shape. This makes it a lot more intuitive to
read.

For apply curve we just had to add the batch dimension into the logic
and change the broadcasting of tensors to account for this new
dimension.

Next we will test and debug these changes by comparing inputs and
outputs of this new code vs. the original code.