Vectorize loss function for CURL model
The loss function had support for batches but it operated on each image
one at a time. Furthermore it made calls to rgb to lab and hsv
conversion methods that only support batches now due to our changes.
Because of this we decided to go ahead and vectorize the loss function
as well. At a high level we did this by removing the for loop, changing
indices to assume there's a batch dim (both in tensors and in
operations), and ensuring the sematics of our vectorized operations
captured the originals.