Refactor Dataset code, add support for different sized images
The previous implementation had batch size 1 which meant that the input
images could vary in size. Now that we have larger batches we need to
convert the images to the same size so we can stack them into a batch.
We added a center crop and a pad operation to make sure that we get
images of the same size. We also got rid of a lot of convoluted,
redundant code in the __getitem__ method for the dataset class.