Refactor model.py to simplify CURLLayer
We found that some of the layer classes like `GlobalPoolingBlock`,
`MaxPoolBlock`, and `Block` were wrappers around simple operators. We
got rid of these classes to simplify the overall implementation of the
CURLLayer. We created a new `ConvStack` layer that contains all of the
convolutions, poolings, etc. that are applied to each channel, which
follow the same pattern. We used this class to wrap and remove a lot of
the code in the forward method of CURLLayer.

We also added a static `convert` method to CURLLayer that captures the
pattern of clamp -> convert color space -> clmap that was repeated in
the forward method of CURLLayer. We replaced the code in the forward
method with this new method to make it more succinct.

The new refactor makes the steps for each color space in the forward
call of CURLLayer clearer: change color space -> stack image and feature
maps -> compute curve parameters -> apply curve. We repeat this pattern
now in a more readable and succinct way.