Refactor data.py to simplify and improve performance
We noticed that all of the `Samsung*` classes in this module were not
used elsewhere. We also found that the Adobe5kDataLoader was really slow
for running the simple task of creating a dict of paths. We figured this
was due to the multiple nested loops in the `load_data` method and using
os.walk. We eliminated the `Samsung*` and `Adobe5kDataLoader` classes,
and created new static methods in the module for creating the data dict
needed for the custom Dataset class. Now the creation of the data
loaders takes seconds instead of 10 minutes for the training data. It's
also much simpler and easier to understand.