Significant changes to modeling module and approach
We've been freestyling a bit for the past few days. We realized that
using only the global portion of CURL, which we called GlobalCURL,
really wasn't working too well. After playing around with changing the
backbone to SegNet (https://arxiv.org/abs/1511.00561), creating a single
layer network for overfitting without much success, playing with the
loss functions, etc., we realized CURL depends a LOT on the
encoder-decoder. This made the model intractible for our application as
we need a resolution independent architecture.

We started approaching the problem as a regression problem. We found
pre-trained backbones and started regressing CURL parameters. That
didn't work too well, but we found that regressing polynomial
coefficients like in DeepLPF does work quite well! After this we
iterated over different backbones and ways of mixing information across
channels. We found that efficientnetv2 is a superb backbone (better than
efficientnet or resnet, by a lot). We also found that combining losses
and residuals from different color spaces leads to qualitatively
significantly better results.

The changes to the model.py file reflect our change in focus, new
modeling approaches, etc. In summary:

- We moved all ssim and mssim calculation to image_processing.py to have
a single implementation of these methods, rather than two like in the
original repo.
- We incorporated masking into the loss function to get better signal
and ignore masking pixels as much as possible.
- We got rid of the gradient regularizer term as CURL is not our main
model anymore and this term didn't do much for us anyways.
- Remove convolutional stacks from CURLLayer, make L, R, and H
parameters for the forward function to be inferred by a better model
(e.g. efficientnetv2).
- Get rid of wrapper `convert` method. It adds unneeded complexity and
made it harder to use CURLLayer in other models downstream.
- Replace GlobalCURLNet with GCURLNet, which uses efficientnetv2 to
regress curve parameters (much more powerful) but then does the regular
CURL we know and love.
- Implement ChannelPolyLayer, which calculates all monomials of a degree
N polynomial with M variables and whose forward method combines the
monomials with their coefficients. This is our (better, more flexible)
version of the cubic layer of DeepLPF.
- Implement PolyRegNet, a net that uses a classifier backbone to produce
polynomial coefficients and applies it to an image in a single color
space.
- Implement TriSpaceRegNet, a net with a classifier backbone to produce
polynomial coefficients and applies them to an image in 3 different
color spaces (rgb, lab, and hsv). We found this to work better, it is
the best of CURL and DeepLPF together! This net normalizes residuals
computed using a sigmoid and centers them to get the correct range for
our predictions. It also adds additional MLPs on top of our backbone for
additional modeling power (works quite well).