Separate reg params & image creation, pass mask separately
In previous iterations we would apply the mask to images in
preprocessing before passing to the model. This was for convenience, but
it's better to pass mask separately so we can work with the image
without mask. This gives us maximal flexibility for our mobile
photography applications. We thus changed the model and loss interfaces
to take masks separately from the image batch.

Additionally, we wanted the option to build the logic to apply the image
transformation in Swift. To that end we separated the regression
parameter estimation from the application of said parameters. This way,
we can pass only the parameters to our iOS program and iOS can handle
the regression given the parameters. This shoudl help with performance
down the line.