Add support for regular training on top of DP and DDP
We found that training in DDP is unstable. We tried DP and although the
training and validation losses are stable we end up running into
https://github.com/pytorch/pytorch/issues/67864 and
https://github.com/facebookresearch/maskrcnn-benchmark/issues/103.
To circumvent all of these problems we just decided to go back to
regular training. It's a shame to not be able to iterate faster by
running quicker experiments but it's taking more time than it's worth
optimizing for parallelization.