Remove size check, width/height params. Add explicit poly layer.
The poly layer we implemented is not compatible with Core ML. Core ML does not support the prod operation nor does it support the pow operation when the powers are a tensor. We thus created a method for generating the equivalent polynomial, term by term. This approach is supported by Core ML. We used this method to generate the appropriate formula and created a separate layer implemeting this formula. This can be passed to the main TriSpaceRegNet instead of ChannelPolyLayer.

The new polylayer, Deg4MobilePolyLayer, has problems. It does not produce the same results as ChannelPolyLayer. We need to debug this problem. This is the last issue we have before getting our mobile compatible model!